pyproject.toml
[tool.poetry]
name = "waitinglist-backend"
version = "0.1.0"
description = "FastAPI backend for the WaitingList project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { version = "^0.115.4", extras = ["standard"] }
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"
slowapi = "^0.1.9"
resend = "^2.15.0"
email-validator = "^2.2.0"
python-multipart = "^0.0.20"
itsdangerous = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[tool.poetry.scripts]
dev = "src.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

